Refactor vm_runner configuration files
