+++ apps/vm_runner/.gitignore
@@ -1,3 +1,4 @@
 service_account.json
 deno.lock
 vm_runner
+prod_urls.ts
